+++ crates/cascette-client-storage/src/shmem/platform_unix.rs
@@ -12,7 +12,7 @@ use std::time::{Duration, Instant};
 use std::{ptr, slice};
 
 use libc::{MAP_SHARED, O_CREAT, O_RDWR, PROT_READ, PROT_WRITE, S_IRUSR, S_IWUSR};
-use libc::{c_void, mode_t, off_t, size_t};
+use libc::{c_uint, c_void, mode_t, off_t, size_t};
 use libc::{close, ftruncate, mmap, munmap, shm_open};
 
 use crate::{Result, StorageError};
@@ -66,7 +66,7 @@ impl PlatformShmem {
             shm_open(
                 c_name.as_ptr(),
                 O_CREAT | O_RDWR,
-                (S_IRUSR | S_IWUSR) as mode_t,
+                (S_IRUSR | S_IWUSR) as mode_t as c_uint,
             )
         };
